refactor: GoToLoginButton으로 컴포넌트 이름을 명확하게 수정
