+++ frontend/src/domains/auth/components/GoToLoginButton/GoToLoginButton.styles.ts
@@ -2,9 +2,9 @@ import { css } from '@emotion/react';
 
 import theme from '@/styles/theme';
 
-const LoginButtonStyle = css`
+const GoToLoginButtonStyle = css`
   width: 50rem;
   background-color: ${theme.colors.white};
   box-shadow: 0 8px 16px rgb(0 0 0 / 26%);
 `;
-export { LoginButtonStyle };
+export { GoToLoginButtonStyle };
